requirements.txt
python-dotenv
google-generativeai==0.8.3
python-telegram-bot==21.6
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.47.0
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
APScheduler==3.10.4
tzlocal==5.2
python-magic==0.4.27
python-magic-bin==0.4.14 ; sys_platform == 'win32'
